package changes
moved all item subclasses to the same package and the world, terrain,
and cell classes to the main package.
