+++ HackISU/bin/.gitignore
@@ -4,3 +4,4 @@
 /oneUseItems/
 /textGame/
 /weapons/
+/items/
