+++ HackISU/src/items/Boots.java
@@ -1,13 +1,16 @@
-package armor;
+package items;
 
-import textGame.Item;
-
-public class Helm extends Armor{
+/**
+ * 
+ * @author Colt Rogness
+ *
+ */
+public class Boots extends Armor{
 	
 	/**
-	 * constructs an empty helm
+	 * constructs an empty boot
 	 */
-	public Helm() {
+	public Boots() {
 		name = "";
 		defense = 0;
 		value = 0;
@@ -16,11 +19,11 @@ public class Helm extends Armor{
 	}
 	
 	/**
-	 * constructs a helm with the given stats
+	 * constructs a boot with the given stats
 	 * @param durability - how many hits it can take
 	 * @param defense - how much damage it removes
 	 */
-	public Helm(String name, int durability, int defense) {
+	public Boots(String name, int durability, int defense) {
 		super.durability = durability;
 		super.defense = defense;
 		super.name = name;
@@ -30,15 +33,14 @@ public class Helm extends Armor{
 
 	@Override
 	public Item clone() {
-		return new Helm(name, durability, defense);
+		return new Boots(name, durability, defense);
 	}
 	
 	/**
 	 * 
 	 * @return - the type of armor this is
 	 */
-	public ArmorType getType() {
-		return ArmorType.HELM;
+	public ArmorType getType(){
+		return ArmorType.BOOTS;
 	}
-	
 }
