+++ HackISU/src/items/Melee.java
@@ -1,4 +1,4 @@
-package weapons;
+package items;
 
 /**
  *
