+++ HackISU/src/items/Ranged.java
@@ -1,4 +1,4 @@
-package weapons;
+package items;
 
 public class Ranged extends Weapon{
