+++ HackISU/src/items/Weapon.java
@@ -1,6 +1,4 @@
-package weapons;
-
-import textGame.Item;
+package items;
 
 /**
  *
