+++ HackISU/src/items/WeaponType.java
@@ -1,4 +1,4 @@
-package weapons;
+package items;
 
 /**
  *
