+++ HackISU/src/textGame/Terrain.java
@@ -1,4 +1,4 @@
-package game;
+package textGame;
 
 /**
  *
